Array

+++ Problems_Set/Sheet 3 - Array/Y_Range_sum.cpp
@@ -7,11 +7,11 @@ int main()
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    int n, q;
+    ll n, q;
     cin >> n >> q;
-    vector<int> a(n+1), b(n+1, 0);
+    vector<ll> a(n+1), b(n+1, 0);
 
-    for (int i = 1; i <= n; i++)
+    for (ll i = 1; i <= n; i++)
     {
         cin >> a[i];
         b[i] = b[i-1] + a[i];
@@ -19,7 +19,7 @@ int main()
 
     while (q--)
     {
-        int l, r;
+        ll l, r;
         cin >> l >> r;
         cout << b[r] - b[l-1] << "\n";
     }
